AVL downloader removing unique database constraint
Removing unique constraint on database `identifier_hash` column.
Hashes are still checked with in memory list of previous hashes and inserts won't be attempted for any already found, this won't prevent duplicates in the database but will reduce the number of rows inserted that have been seen in the current run of the tool.